🐛 fix: Resolve EVM compilation errors with type casting
Fix multiple compilation errors related to std.math.shr and std.math.shl
type mismatches:

- bytecode.zig: Add explicit u3 type casts for shift amounts in bitmap operations
- memory.zig: Use consistent u64 types for memory size calculations
- precompiles.zig: Fix type mismatch in bytesToU32 function

These fixes resolve build failures and ensure proper type safety in
mathematical operations throughout the EVM module.

+++ src/evm/memory.zig
@@ -250,7 +250,7 @@ pub fn Memory(comptime config: MemoryConfig) type {
             const current_size = @as(u64, @intCast(self.size_internal()));
             if (new_size_u64 <= current_size) return 0;
             const new_words = std.math.shr(u64, new_size_u64 + 31, 5); // Using std.math.shr instead of / 32
-            const current_words = std.math.shr(u32, current_size + 31, 5); // Using std.math.shr instead of / 32
+            const current_words = std.math.shr(u64, current_size + 31, 5); // Using std.math.shr instead of / 32
             const new_cost = calculate_memory_cost(new_words);
             const current_cost = calculate_memory_cost(current_words);
             return new_cost - current_cost;
